run python through 2to3

+++ src/gdb/python/lib/gdb/printer/bound_registers.py
@@ -26,7 +26,7 @@ class MpxBound128Printer:
     def to_string(self):
         upper = self.val["ubound"]
         lower = self.val["lbound"]
-        size = (long)((upper) - (lower))
+        size = (int)((upper) - (lower))
         if size > -1:
             size = size + 1
         result = "{{lbound = {}, ubound = {}}} : size {}".format(lower, upper, size)
